requirements.txt
langchain
langchain_groq
langchain_core
langchain_community
flask
flask_cors
python-dotenv
chromadb
langchain-chroma
openai
tiktoken
unstructured
pdfminer
pymupdf
langchain-openai
